Extracting results to its own module
