+++ docs/api/results.md
@@ -1,6 +1,5 @@
 # Results
 
-::: neurotechdevkit.scenarios
+::: neurotechdevkit.results
     options:
       show_root_toc_entry: false
-      filters: ["!Scenario", "!^_[^_]"]
