+++ src/neurotechdevkit/results/__init__.py
@@ -1,0 +1,27 @@
+# noqa: D104
+# preventing package docstring to be rendered in documentation
+from ._results import (
+    PulsedResult,
+    PulsedResult2D,
+    PulsedResult3D,
+    Result,
+    SteadyStateResult,
+    SteadyStateResult2D,
+    SteadyStateResult3D,
+    create_pulsed_result,
+    create_steady_state_result,
+    load_result_from_disk,
+)
+
+__all__ = [
+    "create_steady_state_result",
+    "create_pulsed_result",
+    "load_result_from_disk",
+    "Result",
+    "SteadyStateResult",
+    "SteadyStateResult2D",
+    "SteadyStateResult3D",
+    "PulsedResult",
+    "PulsedResult2D",
+    "PulsedResult3D",
+]
