+++ src/neurotechdevkit/scenarios/__init__.py
@@ -1,35 +1,12 @@
-# noqa: D104
-# preventing package docstring to be rendered in documentation
+"""Scenarios module."""
 from . import materials
 from ._base import Scenario, Scenario2D, Scenario3D
-from ._results import (
-    PulsedResult,
-    PulsedResult2D,
-    PulsedResult3D,
-    Result,
-    SteadyStateResult,
-    SteadyStateResult2D,
-    SteadyStateResult3D,
-    create_pulsed_result,
-    create_steady_state_result,
-    load_result_from_disk,
-)
 from ._scenario_0 import Scenario0
 from ._scenario_1 import Scenario1_2D, Scenario1_3D
 from ._scenario_2 import Scenario2_2D, Scenario2_3D
 
 __all__ = [
-    "create_steady_state_result",
-    "create_pulsed_result",
-    "load_result_from_disk",
     "materials",
-    "Result",
-    "SteadyStateResult",
-    "SteadyStateResult2D",
-    "SteadyStateResult3D",
-    "PulsedResult",
-    "PulsedResult2D",
-    "PulsedResult3D",
     "Scenario",
     "Scenario0",
     "Scenario1_2D",
